Refactor project structure to reduce worker size and group data.
