+++ .gitignore
@@ -1,4 +1,5 @@
 .history
 .DS_Store
+.vscode
 dist
 node_modules
