+++ src/genetic-algorithm/index.js
@@ -1,0 +1,1 @@
+export { default as GeneticAlgorithm } from "./main";
