+++ src/parallel/index.js
@@ -1,0 +1,1 @@
+export { default as Parallel } from "./parallel";
