+++ src/svg-parser/index.js
@@ -1,0 +1,1 @@
+export { default as SvgParser } from "./main";
