+++ src/svg-parser/matrix.js
