+++ src/svg-parser/pathsegpolyfill.js
